shop/requirements.txt
appdata==2.2.1
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored==2.2.4
Django==5.0.6
django-admin-decorators==0.1
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.7.1
django-filter==24.2
django-js-asset==2.2.0
django-render-partial==0.4
djangorestframework==3.15.1
idna==3.7
kavenegar==1.1.2
mysql-connector-python==8.4.0
pillow==10.3.0
requests==2.31.0
sqlparse==0.5.0
to-requirements.txt==2.0.8
tzdata==2024.1
urllib3==2.2.1
